refactor(tests): extract shared mockDB and mockStreamProviderWithSpy utilities
